requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
requests
Pillow